TECHDBT-1236 - Centralize AMA Assets
* Clean up our internal requires for code. This gem contains
  certain code that is only needed when running locally (i.e.
  not mounted as an engine in an application) when performing
  deployments.
* We were getting errors when mounting the gem in a host application
  because this gem's configuration values would override the host
  application.
* To resolve this, move all internal code into the
  lib/ama/styles/internal directory for clarity (and namespace
  the classes accordingly).
* No longer use initializers and environments to set config - instead,
  configuration is set in a ruby file that is only loaded internally.

SEE: https://ama-digital.myjetbrains.com/youtrack/issue/TECHDBT-1236
