--- config/environments/production.rb
@@ -1,62 +1,0 @@
-# frozen_string_literal: true
-
-Rails.application.configure do
-  # Settings specified here will take precedence over those in
-  # config/application.rb.
-
-  # Code is not reloaded between requests.
-  config.cache_classes = true
-
-  # Eager load code on boot. This eager loads most of Rails and
-  # your application in memory, allowing both threaded web servers
-  # and those relying on copy on write to perform better.
-  # Rake tasks automatically ignore this option for performance.
-  config.eager_load = true
-
-  # Full error reports are disabled and caching is turned on.
-  config.consider_all_requests_local       = false
-  config.action_controller.perform_caching = true
-
-  # Enable Rack::Cache to put a simple HTTP cache in front of your application
-  # Add `rack-cache` to your Gemfile before enabling this.
-  # For large-scale production use, consider using a caching reverse proxy like
-  # NGINX, varnish or squid.
-  # config.action_dispatch.rack_cache = true
-
-  # Disable serving static files from the `/public` folder by default since
-  # Apache or NGINX already handles this.
-  config.serve_static_files = ENV['RAILS_SERVE_STATIC_FILES'].present?
-
-  # Do not fallback to assets pipeline if a precompiled asset is missed.
-  config.assets.compile = false
-
-  # Asset digests allow you to set far-future HTTP expiration dates on all
-  # assets, yet still be able to expire them through the digest params.
-  config.assets.digest = true
-
-  # Use the lowest log level to ensure availability of diagnostic information
-  # when problems arise.
-  config.log_level = :debug
-
-  # Enable locale fallbacks for I18n (makes lookups for any locale fall back to
-  # the I18n.default_locale when a translation cannot be found).
-  config.i18n.fallbacks = true
-
-  # Send deprecation notices to registered listeners.
-  config.active_support.deprecation = :notify
-
-  # Use default logging formatter so that PID and timestamp are not suppressed.
-  config.log_formatter = ::Logger::Formatter.new
-
-  # Do not dump schema after migrations.
-  config.active_record.dump_schema_after_migration = false
-
-  config.aws_access_key_id = ENV.fetch('AWS_ACCESS_KEY_ID')
-  config.aws_secret_access_key = ENV.fetch('AWS_SECRET_ACCESS_KEY')
-  config.redis_endpoint = ENV.fetch('REDIS_ENDPOINT')
-  config.api_deployment_url = ENV.fetch('API_DEPLOYMENT_URL')
-  config.api_deployment_user = ENV.fetch('API_DEPLOYMENT_USER')
-  config.api_deployment_password = ENV.fetch('API_DEPLOYMENT_PASSWORD')
-  config.cloudfront_url = ENV.fetch('ASSET_CLOUDFRONT_URL')
-  config.assets_bucket_name = ENV.fetch('S3_ASSETS_BUCKET_NAME')
-end
