--- config/environments/test.rb
@@ -1,44 +1,0 @@
-# frozen_string_literal: true
-
-Rails.application.configure do
-  # Settings specified here will take precedence over those in
-  # config/application.rb.
-
-  # The test environment is used exclusively to run your application's
-  # test suite. You never need to work with it otherwise. Remember that
-  # your test database is "scratch space" for the test suite and is wiped
-  # and recreated between test runs. Don't rely on the data there!
-  config.cache_classes = true
-
-  # Do not eager load code on boot. This avoids loading your whole application
-  # just for the purpose of running a single test. If you are using a tool that
-  # preloads Rails for running tests, you may have to set it to true.
-  config.eager_load = true
-
-  # Configure static file server for tests with Cache-Control for performance.
-  config.public_file_server.enabled = true
-  config.public_file_server.headers = {
-    'Cache-Control' => 'public, max-age=3600'
-  }
-
-  # Show full error reports and disable caching.
-  config.consider_all_requests_local = true
-
-  # Raise exceptions instead of rendering exception templates.
-  config.action_dispatch.show_exceptions = false
-
-  # Randomize the order test cases are executed.
-  config.active_support.test_order = :random
-
-  # Print deprecation notices to the stderr.
-  config.active_support.deprecation = :stderr
-
-  config.aws_access_key_id = 'aws-not-secret'
-  config.aws_secret_access_key = 'super-secret'
-  config.redis_endpoint = 'localhost'
-  config.api_deployment_url = 'http://test.ama.ab.ca/api/v1/assets/deployments.json'
-  config.api_deployment_user = 'api_user'
-  config.api_deployment_password = 'api_password'
-  config.cloudfront_url = 'https://test.cloudfront.net/'
-  config.assets_bucket_name = 'ama_test_assets'
-end
