--- config/initializers/aws.rb
@@ -1,9 +1,0 @@
-# frozen_string_literal: true
-
-Aws.config.update(
-  region: 'us-west-2',
-  credentials: Aws::Credentials.new(
-    Rails.configuration.aws_access_key_id,
-    Rails.configuration.aws_secret_access_key
-  )
-)
