--- config/initializers/request_logging.rb
@@ -1,4 +1,0 @@
-# frozen_string_literal: true
-require 'rest_client/jogger'
-
-RestClient::Jogger::EventSubscriber.new.subscribe
