+++ lib/ama_styles.rb
@@ -9,4 +9,3 @@ require 'dotenv/rails-now'
 require 'ama/styles/base'
 require 'ama/styles/resolver'
 require 'ama/styles/engine'
-require 'ama/styles/deployment'
